backend/requirements.txt
flask
flask-cors
psycopg2-binary
gunicorn
python-dotenv==1.0.1